ai-tutor/tutor-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
PyJWT==2.9.0
httpx==0.27.2
pypdf==5.4.0
